+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ballerine/headless-example
 
+## 0.3.104
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.105
+
+## 0.3.103
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.83
+  - @ballerine/workflow-browser-sdk@0.6.104
+
 ## 0.3.102
 
 ### Patch Changes