+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ballerine/workflows-service
 
+## 0.7.112
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.105
+  - @ballerine/workflow-node-sdk@0.6.105
+
+## 0.7.111
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.83
+  - @ballerine/workflow-core@0.6.104
+  - @ballerine/workflow-node-sdk@0.6.104
+
 ## 0.7.110
 
 ### Patch Changes